pyproject.toml
[tool.poetry]
name = "yandex-eco-fest-bot"
version = "0.1.0"
description = ""
authors = ["arsenyvolodko <arvolodko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.40"
aiogram = "3.5.0"
asyncpg = "^0.30.0"
pydantic = "2.7.0"
python-dotenv = "1.0.1"
greenlet = "^3.2.2"
redis = "^6.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
